pyproject.toml
[project]
name = "bot7"
version = "0.1.0"
description = "Bot7"
authors = [{ name = "Uladzislau Kakhniuk" }]
requires-python = ">=3.12.3"
dependencies = ["requests", "beautifulsoup4", "pyyaml", "openai", "fastapi", "uvicorn", "discord.py", "pydantic"]

[project.scripts]
app = "app.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "app.*", "collector", "collector.*", "generator", "generator.*", "database", "database.*"]
namespaces = false
